Makefile: Use installed gcc-with-cpychecker
The gcc-with-cpychecker script finds python.so using relative to itself:
...
abspath = os.path.abspath(os.path.dirname(sys.argv[0]))
PLUGIN = os.path.join(abspath, 'python.so')
...

Consequently, when using $(srcdir)./gcc-with-cpychecker in a seperate build
dir, it cannot find python.so.

Create an installed gcc-with-cpychecker by copying it from the source dir, and
use it to run the demo.